requirements.txt
google-generativeai==0.8.3
requests==2.31.0
python-dotenv==1.0.0
